set up back end api call to fetch all events
