+++ client/src/containers/App.js
@@ -13,6 +13,7 @@ import AboutPage from "../components/AboutPage";
 class App extends Component {
   componentDidMount() {
     this.props.fetchUser();
+    this.props.fetchAllEvents();
   }
 
   render() {
